pyproject.toml
[project]
name = "organization-information-fetcher"
version = "0.1.0"
description = "Project to retrieve information about an organization"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "crewai>=0.105.0",
    "dateparser>=1.2.1",
    "google-search-results>=2.4.2",
    "googlesearch-python>=1.3.0",
    "langchain>=0.3.19",
    "langchain-cli>=0.0.35",
    "langchain-community>=0.3.18",
    "langchain-mistralai>=0.2.7",
    "mistralai>=1.5.1",
    "pandas>=2.2.3",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.6.1",
    "sentence-transformers>=3.4.1",
    "streamable>=1.4.8",
]

[project.scripts]
organization-information-fetcher = "organization_information_fetcher_app:main"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]
